struct(T10): split render.rs into render/{mod,setup,layout,panels,modals,game_over}.rs
