+++ src/ui/input/mod.rs
@@ -1,3 +1,5 @@
+//! Keyboard input event handlers for each screen family, implementing `App` input helpers.
+
 mod deals;
 mod file;
 mod main;
